delete Game and Player
