+++ BlazorRummiSolve/Components/Shared/PlayerComponent.razor
@@ -9,7 +9,7 @@
 </div>
 
 @code {
-    [Parameter] public required SimplePlayer Player { get; set; }
+    [Parameter] public required Player Player { get; set; }
     [Parameter] public bool ShowHint { get; set; }
     [Parameter] public required Set CurrentPlayerRack { get; set; }
 }
