+++ BlazorRummiSolve/Components/Shared/PlayersContainerComponent.razor
@@ -17,9 +17,9 @@
 </div>
 
 @code {
-    [Parameter] public required SimplePlayer CurrentPlayer { get; set; }
+    [Parameter] public required Player CurrentPlayer { get; set; }
 
-    [Parameter] public required List<SimplePlayer> OtherPlayers { get; set; }
+    [Parameter] public required List<Player> OtherPlayers { get; set; }
 
     [Parameter] public bool ShowHint { get; set; }
