+++ RummiSolve/RummiSolve/Program.cs
@@ -5,7 +5,5 @@ public static class Program
     private static void Main()
     {
         //BenchmarkRunner.Run<RummiBench>();
-        RummiBench.TestMultiPlayerGame();
     }
-    
 }
