+++ RummiSolve/RummiSolve/RummiBench.cs
@@ -1,4 +1,3 @@
-using System.Diagnostics;
 using BenchmarkDotNet.Attributes;
 using RummiSolve.Solver.Combinations.First;
 using RummiSolve.Solver.Incremental;
@@ -18,16 +17,10 @@ public class RummiBench
         _game.InitializeGame(_playerNames);
     }
 
-    [Benchmark]
-    public void TestMultiPlayerGame_Benchmark()
-    {
-        _game.Start();
-    }
-
 
     public static void TestSimpleGame()
     {
-        SimpleGame game = new(Guid.Parse("8f53d490-db85-4962-8886-8a49c0e2afb8"));
+        Game game = new(Guid.Parse("8f53d490-db85-4962-8886-8a49c0e2afb8"));
 
         var listNames = new List<string> { "Antoine", "Matthieu", "Maguy" };
         game.InitializeGame(listNames);
@@ -62,7 +55,7 @@ public class RummiBench
 
     public static void TestSimpleGame2()
     {
-        var game = new SimpleGame(Guid.Parse("8f53d490-db85-4962-8886-8a49c0e2afb8"));
+        var game = new Game(Guid.Parse("8f53d490-db85-4962-8886-8a49c0e2afb8"));
         var playerNames = new List<string> { "Antoine", "Matthieu", "Maguy" };
 
         game.InitializeGame(playerNames);
@@ -97,19 +90,6 @@ public class RummiBench
         }
 
         Console.WriteLine("=== FIN DE LA PARTIE ===");
-    }
-
-    public static void TestMultiPlayerGame()
-    {
-        Game game = new(Guid.Parse("8ae34041-70c4-4338-9bb5-1c68bed2cfdc"));
-        //Guid.Parse("b3b4ff72-9e5d-4c3b-a0bf-b08e193fb156") Guid.Parse("74cdccda-9261-460c-9414-31d7270ad2a1")
-
-        var listNames = new List<string> { "Antoine", "Matthieu", "Maguy" };
-        game.InitializeGame(listNames);
-        var gameStopwatch = Stopwatch.StartNew();
-        game.Start();
-        gameStopwatch.Stop();
-        Console.WriteLine($"Game duration: {gameStopwatch.Elapsed.TotalSeconds} seconds");
     }
